This commit was manufactured by cvs2svn to create tag 'tulip_2_0_5'.
git-svn-id: https://svn.code.sf.net/p/auber/code/tags/tulip_2_0_5@435 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/include/tulip/ObservableGraph.h
@@ -31,7 +31,7 @@ namespace std {
   template <>
   struct less<GraphObserver *> {
     size_t operator()(const GraphObserver * obs1,const GraphObserver *obs2) const {
-      return (unsigned)obs1<(unsigned)obs2;
+      return (unsigned long)obs1<(unsigned long)obs2;
     }
   };
 }
